% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descending.R
\name{descending}
\alias{descending}
\title{Reverse a sorting function}
\usage{
descending(f)
}
\arguments{
\item{f}{a function to modify.}
}
\value{
A function returning a numeric vector which, if passed to
\code{\link{order}}, will be used to sort the original data.
}
\description{
Transform a function (typically used in \code{\link{sort_by}}), so that its
ouput can be sorted in descending order.
}
\examples{
desc_abs <- descending(abs)

x <- -2:1
order(abs(x))
order(desc_abs(x))
}
